Require version number and name outputs according to target, version and date

+++ build.sh
@@ -1,10 +1,23 @@
+if [ "$#" -ne 1 ]; then
+  echo "Usage: $0 GSTREAMER_VERSION" >&2
+  exit 1
+fi
+
+if [[ -z "${GSTREAMER_ROOT_ANDROID}" ]]; then
+  echo "You must define an environment variable called GSTREAMER_ROOT_ANDROID and point it to the folder where you extracted the GStreamer binaries"
+  exit 1
+fi
+
+VERSION=$1
+DATE=`date "+%Y%m%d-%H%M%S"`
+
 rm -rf out
 mkdir out
 
 for TARGET in armv7 x86
 do
   NDK_APPLICATION_MK="jni/${TARGET}.mk"
-  echo "Building for target ${TARGET} with ${NDK_APPLICATION_MK}"
+  echo "\n\n=== Building GStreamer ${VERSION} for target ${TARGET} with ${NDK_APPLICATION_MK} ==="
 
   ndk-build NDK_APPLICATION_MK=$NDK_APPLICATION_MK
 
@@ -27,7 +40,8 @@ do
   sed -i -e 's?Libs.private.*?Libs.private: -lgstreamer_android?g' pkgconfig/*
   rm -rf pkgconfig/*-e*
   cd ..
-  zip -v out/$GST_LIB.zip $GST_LIB/* -r
+  ZIP="gstreamer-${LIB}-${VERSION}-${DATE}"
+  zip -v out/$ZIP.zip $GST_LIB/* -r
   rm -rf $GST_LIB
 done
 
@@ -35,3 +49,5 @@ rm -rf libs obj
 
 zip -v out/src.zip src -r
 rm -rf src
+
+echo "\n*** Done ***\n`ls out`"
